Removing old playerserver code
