+++ PlayerServer/main.go
@@ -15,9 +15,13 @@ func main() {
 		log.Fatalf("problem opening %s %v", dbFileName, err)
 	}
 
-	store := NewFileSystemPlayerStore(db)
-	server := NewPlayerServer(store)
+	store, err := NewFileSystemPlayerStore(db)
 
+	if err != nil {
+		log.Fatalf("problem creating file system player store, %v", err)
+	}
+
+	server := NewPlayerServer(store)
 	if err := http.ListenAndServe(":5000", server); err != nil {
 		log.Fatalf("could not listen on port 5000 %v", err)
 	}
