+++ PlayerServer/server_integration_test.go
@@ -6,10 +6,13 @@ import (
 	"testing"
 )
 
-func TestRecordingWinsAndRetrievingThem(t *testing.T) {
-	database, cleanDatabase := createTempFile(t, "")
+func TestRecordingWinsAndGettingThem(t *testing.T) {
+	database, cleanDatabase := createTempFile(t, `[]`)
 	defer cleanDatabase()
-	store := NewFileSystemPlayerStore(database)
+	store, err := NewFileSystemPlayerStore(database)
+
+	assertNoError(t, err)
+
 	server := NewPlayerServer(store)
 	player := "Pepper"
 
@@ -18,22 +21,21 @@ func TestRecordingWinsAndRetrievingThem(t *testing.T) {
 	server.ServeHTTP(httptest.NewRecorder(), newPostWinRequest(player))
 
 	t.Run("get score", func(t *testing.T) {
-		response := httptest.NewRecorder()
-		server.ServeHTTP(response, newGetScoreRequest(player))
-		assertResponseCode(t, response.Code, http.StatusOK)
-
-		assertResponseBody(t, response.Body.String(), "3")
+		resp := httptest.NewRecorder()
+		server.ServeHTTP(resp, newGetScoreRequest(player))
+		assertResponseCode(t, resp.Code, http.StatusOK)
+		assertResponseBody(t, resp.Body.String(), "3")
 	})
 
 	t.Run("get league", func(t *testing.T) {
-		response := httptest.NewRecorder()
-		server.ServeHTTP(response, newLeagueRequest())
-		assertResponseCode(t, response.Code, http.StatusOK)
-
-		got := getLeagueFromResponse(t, response.Body)
+		resp := httptest.NewRecorder()
+		server.ServeHTTP(resp, newLeagueRequest())
+		assertResponseCode(t, resp.Code, http.StatusOK)
+		got := getLeagueFromResponse(t, resp.Body)
 		want := []Player{
 			{"Pepper", 3},
 		}
 		assertLeague(t, got, want)
 	})
+
 }
