--- PlayerServer2/FileSystemStore.go
@@ -1,81 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-	"sort"
-)
-
-// FileSystemPlayerStore is a PlayerStore implementation using the filesystem
-type FileSystemPlayerStore struct {
-	database *json.Encoder
-	league   League
-}
-
-func initialisePlayerDBFile(file *os.File) error {
-	file.Seek(0, 0)
-
-	info, err := file.Stat()
-
-	if err != nil {
-		return fmt.Errorf("problem getting file info from file %s, %v", file.Name(), err)
-	}
-
-	if info.Size() == 0 {
-		file.Write([]byte("[]"))
-		file.Seek(0, 0)
-	}
-
-	return nil
-}
-
-// NewFileSystemPlayerStore creates a new Store, allowing for only one nevessary read
-func NewFileSystemPlayerStore(file *os.File) (*FileSystemPlayerStore, error) {
-
-	err := initialisePlayerDBFile(file)
-
-	if err != nil {
-		return nil, fmt.Errorf("problem initialising player db file, %v", err)
-	}
-
-	league, err := NewLeague(file)
-
-	if err != nil {
-		return nil, fmt.Errorf("problem loading player store from file %s, %v", file.Name(), err)
-	}
-
-	return &FileSystemPlayerStore{
-		database: json.NewEncoder(&tape{file}),
-		league:   league,
-	}, nil
-}
-
-// GetLeague reads data from the database, ensuring to go back to the beginning
-func (f *FileSystemPlayerStore) GetLeague() League {
-	sort.Slice(f.league, func(i, j int) bool {
-		return f.league[i].Wins > f.league[j].Wins
-	})
-	return f.league
-}
-
-// GetPlayerScore returns the score of the requested player
-func (f *FileSystemPlayerStore) GetPlayerScore(name string) int {
-	player := f.league.Find(name)
-	if player != nil {
-		return player.Wins
-	}
-	return 0
-}
-
-// RecordWin increments number of wins for requested player
-func (f *FileSystemPlayerStore) RecordWin(name string) {
-	player := f.league.Find(name)
-
-	if player != nil {
-		player.Wins++
-	} else {
-		f.league = append(f.league, Player{name, 1})
-	}
-	f.database.Encode(f.league)
-}
