--- PlayerServer2/FileSystemStore_test.go
@@ -1,125 +1,0 @@
-package main
-
-import (
-	"io/ioutil"
-	"os"
-	"testing"
-)
-
-func TestFileSystemStore(t *testing.T) {
-	t.Run("leeague sorted", func(t *testing.T) {
-		database, cleanDatabase := createTempFile(t, `[
-			{"Name": "Cleo", "Wins": 10},
-			{"Name": "Chris", "Wins": 30}]`)
-		defer cleanDatabase()
-
-		store, err := NewFileSystemPlayerStore(database)
-
-		assertNoError(t, err)
-
-		got := store.GetLeague()
-
-		want := []Player{
-			{"Chris", 30},
-			{"Cleo", 10},
-		}
-
-		assertLeague(t, got, want)
-
-		// read again
-		got = store.GetLeague()
-		assertLeague(t, got, want)
-	})
-
-	t.Run("get player score", func(t *testing.T) {
-		database, cleanDatabase := createTempFile(t, `[
-			{"Name": "Cleo", "Wins": 10},
-			{"Name": "Chris", "Wins": 33}]`)
-		defer cleanDatabase()
-
-		store, err := NewFileSystemPlayerStore(database)
-
-		assertNoError(t, err)
-
-		got := store.GetPlayerScore("Chris")
-
-		want := 33
-		assertScoreEquals(t, got, want)
-	})
-
-	t.Run("returns 0 for untracked player", func(t *testing.T) {
-		database, cleanDatabase := createTempFile(t, `[
-			{"Name": "Cleo", "Wins": 10},
-			{"Name": "Chris", "Wins": 33}]`)
-		defer cleanDatabase()
-
-		store, err := NewFileSystemPlayerStore(database)
-
-		assertNoError(t, err)
-
-		got := store.GetPlayerScore("Tom")
-
-		want := 0
-		assertScoreEquals(t, got, want)
-	})
-
-	t.Run("store wins for existing players", func(t *testing.T) {
-		database, cleanDatabase := createTempFile(t, `[
-			{"Name": "Cleo", "Wins": 10},
-			{"Name": "Chris", "Wins": 30}]`)
-		defer cleanDatabase()
-
-		store, err := NewFileSystemPlayerStore(database)
-
-		assertNoError(t, err)
-
-		store.RecordWin("Chris")
-
-		got := store.GetPlayerScore("Chris")
-		want := 31
-		assertScoreEquals(t, got, want)
-	})
-
-	t.Run("store wins for new players", func(t *testing.T) {
-		database, cleanDatabase := createTempFile(t, `[
-			{"Name": "Cleo", "Wins": 10},
-			{"Name": "Chris", "Wins": 30}]`)
-		defer cleanDatabase()
-
-		store, err := NewFileSystemPlayerStore(database)
-
-		assertNoError(t, err)
-
-		store.RecordWin("Pepper")
-
-		got := store.GetPlayerScore("Pepper")
-		want := 1
-		assertScoreEquals(t, got, want)
-	})
-
-	t.Run("works with an empty file", func(t *testing.T) {
-		database, cleanDatabase := createTempFile(t, "")
-		defer cleanDatabase()
-
-		_, err := NewFileSystemPlayerStore(database)
-
-		assertNoError(t, err)
-	})
-}
-
-func createTempFile(t *testing.T, initialData string) (*os.File, func()) {
-	t.Helper()
-	tempfile, err := ioutil.TempFile("", "db")
-
-	if err != nil {
-		t.Fatalf("could not create temp file %v", err)
-	}
-
-	tempfile.Write([]byte(initialData))
-
-	removeFile := func() {
-		os.Remove(tempfile.Name())
-	}
-
-	return tempfile, removeFile
-}
