--- PlayerServer2/server.go
@@ -1,74 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-)
-
-const jsonContentType = "application/json"
-
-// Player stores a name with a number of wins
-type Player struct {
-	Name string
-	Wins int
-}
-
-// PlayerStore stores score information about players
-type PlayerStore interface {
-	GetPlayerScore(name string) int
-	RecordWin(name string)
-	GetLeague() League
-}
-
-// PlayerServer is a HTTP interface for player information
-type PlayerServer struct {
-	store PlayerStore
-	http.Handler
-}
-
-// NewPlayerServer creats a PlayerServer with routing configured
-func NewPlayerServer(store PlayerStore) *PlayerServer {
-	p := new(PlayerServer)
-
-	p.store = store
-
-	router := http.NewServeMux()
-	router.Handle("/league", http.HandlerFunc(p.leagueHandler))
-	router.Handle("/players/", http.HandlerFunc(p.playerHandler))
-
-	p.Handler = router
-	return p
-}
-
-func (p *PlayerServer) leagueHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", jsonContentType)
-	json.NewEncoder(w).Encode(p.store.GetLeague())
-	w.WriteHeader(http.StatusOK)
-}
-
-func (p *PlayerServer) playerHandler(w http.ResponseWriter, r *http.Request) {
-	player := r.URL.Path[len("/players/"):]
-
-	switch r.Method {
-	case http.MethodPost:
-		p.processWin(w, player)
-	case http.MethodGet:
-		p.showScore(w, player)
-	}
-}
-
-func (p *PlayerServer) processWin(w http.ResponseWriter, player string) {
-	p.store.RecordWin(player)
-	w.WriteHeader(http.StatusAccepted)
-}
-
-func (p *PlayerServer) showScore(w http.ResponseWriter, player string) {
-	score := p.store.GetPlayerScore(player)
-
-	if score == 0 {
-		w.WriteHeader(http.StatusNotFound)
-	}
-
-	fmt.Fprint(w, score)
-}
