--- PlayerServer2/tape_test.go
@@ -1,25 +1,0 @@
-package main
-
-import (
-	"io/ioutil"
-	"testing"
-)
-
-func TestTape_Write(t *testing.T) {
-	file, clean := createTempFile(t, "12345")
-	defer clean()
-
-	tape := &tape{file}
-
-	tape.Write([]byte("abc"))
-
-	file.Seek(0, 0)
-	newFileContents, _ := ioutil.ReadAll(file)
-
-	got := string(newFileContents)
-	want := "abc"
-
-	if got != want {
-		t.Errorf("got '%s', wanted '%s'", got, want)
-	}
-}
